+++ e2e/react-router/src/routes/root.tsx
@@ -24,6 +24,7 @@ const client = init(
 				},
 			}),
 			new SessionReplay('1', {
+				debug: { clientInteractions: true, domRecording: true },
 				privacySetting: 'none',
 				serviceName: 'ryan-test',
 				backendUrl: 'https://pub.observability.ld-stg.launchdarkly.com',
@@ -36,6 +37,7 @@ export default function Root() {
 	const fillColor = 'lightblue'
 	const canvasRef = useRef<HTMLCanvasElement>(null)
 	const [flags, setFlags] = useState<string>()
+	const [session, setSession] = useState<string>()
 
 	useEffect(() => {
 		const canvas = canvasRef.current
@@ -46,14 +48,24 @@ export default function Root() {
 		ctx.fillRect(0, 0, canvas.width, canvas.height)
 	}, [fillColor])
 
+	useEffect(() => {
+		const int = setInterval(() => {
+			setSession(LDRecord.getSession()?.url)
+		}, 1000)
+		return () => {
+			clearInterval(int)
+		}
+	}, [])
+
 	return (
 		<div id="sidebar">
 			<h1>Hello, world</h1>
 			<p>{flags}</p>
+			<p>{session}</p>
 			<canvas width="100" height="100" ref={canvasRef}></canvas>
 			<button
 				onClick={() => {
-					LDObserve.recordLog('hello', 'INFO')
+					LDObserve.recordLog('hello', 'info')
 				}}
 			>
 				LDObserve.recordLog
@@ -73,6 +85,17 @@ export default function Root() {
 				}}
 			>
 				LDRecord.snapshot
+			</button>
+			<button
+				onClick={async () => {
+					setFlags(
+						JSON.stringify(
+							client.variation('enable-session-card-style'),
+						),
+					)
+				}}
+			>
+				client.eval
 			</button>
 			<button
 				onClick={async () => {